Added custom mint function, need tests
I've added the custom mint function, with the hashed message check. It needs to be tests.
I've also added some basic tests for the notary, just to use the Foundry framework a bit and learn how to setup a new test file.
I made a basic Makefile to run my scripts.